refactor: Rename Connection to WebSocketConnection
